requirements.txt
django
djangorestframework
requests
gunicorn
psycopg2
